backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pymysql==1.1.1
pydantic==2.11.7
pydantic-settings==2.10.1
httpx==0.28.1
cryptography==45.0.7
python-multipart==0.0.20
Pillow>=12.1.0,<13.0.0
